chore(virtctl): Move add volume command to own file
